+++ src/test/scala/com/twitter/finagle/postgres/integration/IntegrationSpec.scala
@@ -109,7 +109,7 @@ class IntegrationSpec extends Spec {
         cleanDb(client)
         insertSampleData(client)
 
-        val selectQuery = client.select(
+        val selectQuery = client.selectToSeq(
           "SELECT * FROM %s WHERE str_field='hello' ORDER BY timestamp_field".format(IntegrationSpec.pgTestTable)
         )(
           identity)
@@ -130,7 +130,6 @@ class IntegrationSpec extends Spec {
         firstRow.getOption[Boolean]("bool_field") must equal(Some(false))
         firstRow.getOption[String]("bad_column") must equal(None)
 
-
       }
 
       "execute a select that returns nothing" in {
@@ -140,7 +139,7 @@ class IntegrationSpec extends Spec {
         insertSampleData(client)
 
         val
-        selectQuery = client.select(
+        selectQuery = client.selectToSeq(
           "SELECT * FROM %s WHERE str_field='xxxx' ORDER BY timestamp_field".
             format(
 
@@ -171,7 +170,7 @@ class IntegrationSpec extends Spec {
 
         response must equal(OK(1))
 
-        val selectQuery = client.select(
+        val selectQuery = client.selectToSeq(
 
           "SELECT * FROM %s WHERE str_field='hello_updated'".format(IntegrationSpec.pgTestTable)
         )(
@@ -200,7 +199,7 @@ class IntegrationSpec extends Spec {
 
         response must equal(OK(3))
 
-        val selectQuery = client.select(
+        val selectQuery = client.selectToSeq(
           "SELECT * FROM %s".format(IntegrationSpec.pgTestTable)
         )(identity)
 
@@ -246,7 +245,7 @@ class IntegrationSpec extends Spec {
     
         val numRows = Await.result(preparedQuery)
     
-        val resultRows = Await.result(client.select(
+        val resultRows = Await.result(client.selectToSeq(
           "SELECT * from %s WHERE str_field = 'hello_updated' AND int_field = 4567".format(IntegrationSpec.pgTestTable)
         )(identity))
 
@@ -266,7 +265,7 @@ class IntegrationSpec extends Spec {
 
         val numRows = Await.result(preparedQuery)
 
-        val resultRows = Await.result(client.select(
+        val resultRows = Await.result(client.selectToSeq(
           "SELECT * from %s WHERE str_field = 'hello_updated_some' AND int_field = 4567".format(IntegrationSpec.pgTestTable)
         )(identity))
 
@@ -286,7 +285,7 @@ class IntegrationSpec extends Spec {
 
         val numRows = Await.result(preparedQuery)
 
-        val resultRows = Await.result(client.select(
+        val resultRows = Await.result(client.selectToSeq(
           "SELECT * from %s WHERE str_field IS NULL AND int_field = 4567".format(IntegrationSpec.pgTestTable)
         )(identity))
 
@@ -387,7 +386,7 @@ class IntegrationSpec extends Spec {
           val client = getClient
           cleanDb(client)
 
-          val selectQuery = client.select(
+          val selectQuery = client.selectToSeq(
             "SELECT * FROM %s WHERE unknown_column='hello_updated'".format(IntegrationSpec.pgTestTable)
           )(identity)